// types/TowerTags.ts
export enum TowerTag {
    // Basic
    HasTelescope = "hasTelescope",
    HasObservationBoards = "hasObservationBoards",
    IsNearTouristGuide = "isNearTouristGuide",

    // Accessibility
    HasParking = "hasParking",
    IsNearPublicTransport = "isNearPublicTransport",
    NeedToBorrowKey = "needToBorrowKey",
    NeedToBookVisit = "needToBookVisit",
    SuitableForCyclists = "suitableForCyclists",
    WheelchairAccessible = "wheelchairAccessible",

    // Services
    HasToilet = "hasToilet",
    HasRestaurant = "hasRestaurant",
    HasSnacks = "hasSnacks",
    HasWifi = "hasWifi",
    CanPayByCard = "canPayByCard",
    HasShelter = "hasShelter",
    HasBikeRepairStation = "hasBikeRepairStation",
    HasElectricCharger = "hasElectricCharger",

    // Safety
    HasSteepStairs = "hasSteepStairs",
    HasSmallRailings = "hasSmallRailings",
    HasSlipperySurface = "hasSlipperySurface",
    HasLadder = "hasLadder",
}